Refactor GbfsLoader class
